More 64bit safety fixes

+++ client/snd_mix.c
@@ -101,7 +101,7 @@ LClampDone2:
 //#endif
 #endif
 
-void S_TransferStereo16 (unsigned long *pbuf, int endtime)
+void S_TransferStereo16 (unsigned *pbuf, int endtime)
 {
 	int		lpos;
 	int		lpaintedtime;
@@ -144,9 +144,9 @@ void S_TransferPaintBuffer(int endtime)
 	int 	*p;
 	int 	step;
 	int		val;
-	unsigned long *pbuf;
+	unsigned *pbuf;
 
-	pbuf = (unsigned long *)dma.buffer;
+	pbuf = (unsigned *)dma.buffer;
 
 	if (s_testsound->value)
 	{
